Попытка разделить и властвовать
